ProductFacade 생성 및 API 연동

+++ src/main/java/kr/hhplus/be/server/domain/product/dto/ProductListDto.java
@@ -6,7 +6,7 @@ public record ProductListDto(
         Long id,
         String name,
         double price,
-        long stockQuantity
+        int stockQuantity
 ) {
     public static ProductListDto of(Product product, StockDto stock) {
         return new ProductListDto(
